+++ bundle/src/main/java/com/adobe/acs/commons/redirects/models/ExportColumn.java
@@ -36,7 +36,8 @@ public enum ExportColumn {
     CREATED_BY("Created By", RedirectRule.CREATED_BY_PROPERTY_NAME, String.class, false),
     MODIFIED("Modified", RedirectRule.MODIFIED_PROPERTY_NAME, Calendar.class, false),
     MODIFIED_BY("Modified By", RedirectRule.MODIFIED_BY_PROPERTY_NAME, String.class, false),
-    CACHE_CONTROL("Cache-Control", RedirectRule.CACHE_CONTROL_HEADER_NAME, String.class, false);
+    CACHE_CONTROL("Cache-Control", RedirectRule.CACHE_CONTROL_HEADER_NAME, String.class, false),
+    PRESERVE_QUERY_STRING("Query String", RedirectRule.PRESERVE_QUERY_STRING, String.class, true);
 
     private final String title;
     private final String propertyName;